manually build the json manifest and index files
This needs to be updated to use a proper class for each type of
blob - there are more interlinked pieces than I had thought.

This is not yet working - one of the sha256 checks on the new
tar file is failing for reasons I do not yet understand.
(i.e. sha256sum on the uncompressed contents reports the correct
value)

Signed-off-by: Serge Hallyn <shallyn@cisco.com>